# Cursor/inspect results of ggml_interface.cpp