--- a/nntrainer/tensor/cpu_backend/ggml_interface/ggml_interface.cpp
+++ b/nntrainer/tensor/cpu_backend/ggml_interface/ggml_interface.cpp
@@ -100,29 +100,40 @@ void __ggml_q4_0_8x8_q8_0_GEMM(const unsigned int M, const unsigned int N,
                                const unsigned int lda, const void *B,
                                const unsigned int ldb, float *C,
                                const unsigned int ldc) {
-  // auto &bspool = ThreadPoolManager::getInstance();
+  auto &bspool = ThreadPoolManager::getInstance();
 
   if (M == 1) { // GEMV
-    int n_threads = 4;
+    // Adaptive thread count based on problem size
+    const unsigned int min_work_per_thread = 64;
+    const unsigned int n_threads = std::min(4u, std::max(1u, N / min_work_per_thread));
     unsigned int B_step = sizeof(block_q4_0) * (K / QK4_0);
     unsigned int blocks_per_row = (K + QK8_0 - 1) / QK8_0;
     unsigned int qa_size = sizeof(block_q8_0) * blocks_per_row;
     std::vector<char> QA = std::vector<char>(qa_size);
     ::quantize_row_q8_0(A, QA.data(), K);
 
-#pragma omp parallel for num_threads(n_threads)
-    for (int thread_idx = 0; thread_idx < n_threads; ++thread_idx) {
-      unsigned int M_step_start = (thread_idx * N) / n_threads;     // = 0
-      unsigned int M_step_end = ((thread_idx + 1) * N) / n_threads; // ne01 = N
-
-      M_step_start = (M_step_start % 8) ? M_step_start + 8 - (M_step_start % 8)
-                                        : M_step_start;
-      M_step_end =
-        (M_step_end % 8) ? M_step_end + 8 - (M_step_end % 8) : M_step_end;
-
-      ::ggml_gemv_q4_0_8x8_q8_0(K, (float *)((C) + M_step_start), N,
-                                (void *)((char *)B + M_step_start * B_step),
-                                QA.data(), M, M_step_end - M_step_start);
+    // Use BS thread pool for parallel GEMV
+    std::vector<std::future<void>> futures;
+    futures.reserve(n_threads);
+    
+    for (unsigned int thread_idx = 0; thread_idx < n_threads; ++thread_idx) {
+      futures.emplace_back(bspool.submit_task([=, &QA, &B, &C]() {
+        unsigned int M_step_start = (thread_idx * N) / n_threads;
+        unsigned int M_step_end = ((thread_idx + 1) * N) / n_threads;
+        
+        // Cache-line alignment for better performance
+        M_step_start = (M_step_start % 8) ? M_step_start + 8 - (M_step_start % 8) : M_step_start;
+        M_step_end = (M_step_end % 8) ? M_step_end + 8 - (M_step_end % 8) : M_step_end;
+        
+        ::ggml_gemv_q4_0_8x8_q8_0(K, (float *)((C) + M_step_start), N,
+                                  (void *)((char *)B + M_step_start * B_step),
+                                  QA.data(), M, M_step_end - M_step_start);
+      }));
+    }
+    
+    // Wait for all tasks to complete
+    for (auto &future : futures) {
+      future.wait();
     }
   } else { // GEMM
     unsigned int blocks_per_4_rows = (K + QK8_0 - 1) / QK8_0;
@@ -135,7 +146,6 @@ void __ggml_q4_0_8x8_q8_0_GEMM(const unsigned int M, const unsigned int N,
     // Quantization of activations
     /// @note Heuristic inspection conducted that applying multithreading on
     /// run-time quantization hurts model latency
-    // #pragma omp parallel for collapse(1) num_threads(16)
     for (int i = 0; i < static_cast<int>(M4); i++) {
       ::ggml_quantize_mat_q8_0_4x8(A + 4 * i * K,
                                    QA.data() + i * qa_4_rows_size, K);
@@ -144,18 +154,19 @@ void __ggml_q4_0_8x8_q8_0_GEMM(const unsigned int M, const unsigned int N,
     int step_N = N / delta;
     int step_C = delta;
     int step_B = blocks_per_4_rows * sizeof(block_q4_0) * delta;
-#pragma omp parallel for collapse(1) num_threads(16)
-    for (int i = 0; i < step_N; i++) {
-      ::ggml_gemm_q4_0_8x8_q8_0(K, C + i * step_C, ldc, (char *)B + i * step_B,
-                                QA.data(), M, delta);
-    }
-    /**
-    @todo Add BS threadpool multithread strategy
-    BS::multi_future<void> multi_future = bspool.submit_loop(0, step_N, [&](int
-    i){::ggml_gemm_q4_0_8x8_q8_0(K, C + i * step_C, ldc, (char *)B + i * step_B,
-                                QA.data(), M, delta);});
-      multi_future.wait();
-     */
+    
+    // Adaptive thread count for GEMM
+    const unsigned int optimal_threads = std::min(16u, 
+      std::max(1u, static_cast<unsigned int>(step_N)));
+    
+    // Use BS thread pool for parallel GEMM
+    BS::multi_future<void> multi_future = bspool.submit_loop(0, step_N, 
+      [&](int i) {
+        ::ggml_gemm_q4_0_8x8_q8_0(K, C + i * step_C, ldc, 
+                                  (char *)B + i * step_B,
+                                  QA.data(), M, delta);
+      });
+    multi_future.wait();
   }
 }
 
@@ -164,8 +175,10 @@ void __ggml_q4_K_8x8_q8_K_GEMM(const unsigned int M, const unsigned int N,
                                const unsigned int lda, const void *B,
                                const unsigned int ldb, float *C,
                                const unsigned int ldc) {
+  auto &bspool = ThreadPoolManager::getInstance();
+  
   if (M == 1) { // GEMV
-    int n_threads = 4;
+    const unsigned int n_threads = std::min(4u, std::max(1u, N / 64));
     unsigned int blocks_per_row = (K + QK_K - 1) / QK_K;
     unsigned int qa_size = sizeof(block_q8_K) * blocks_per_row;
     unsigned int B_step = sizeof(block_q4_K) * (K / QK_K);
@@ -174,22 +187,31 @@ void __ggml_q4_K_8x8_q8_K_GEMM(const unsigned int M, const unsigned int N,
 
     ::quantize_row_q8_K(A, QA.data(), K);
 
-#pragma omp parallel for num_threads(n_threads)
-    for (int thread_idx = 0; thread_idx < n_threads; ++thread_idx) {
-      unsigned int M_step_start = (thread_idx * N) / n_threads;     // = 0
-      unsigned int M_step_end = ((thread_idx + 1) * N) / n_threads; // ne01 = N
-
-      M_step_start = (M_step_start % 8) ? M_step_start + 8 - (M_step_start % 8)
-                                        : M_step_start;
-      M_step_end =
-        (M_step_end % 8) ? M_step_end + 8 - (M_step_end % 8) : M_step_end;
-
-      ::ggml_gemv_q4_K_8x8_q8_K(K, (float *)((C) + M_step_start), N,
-                                (void *)((char *)B + M_step_start * B_step),
-                                QA.data(), M, M_step_end - M_step_start);
+    // Use BS thread pool for parallel GEMV  
+    std::vector<std::future<void>> futures;
+    futures.reserve(n_threads);
+    
+    for (unsigned int thread_idx = 0; thread_idx < n_threads; ++thread_idx) {
+      futures.emplace_back(bspool.submit_task([=, &QA, &B, &C]() {
+        unsigned int M_step_start = (thread_idx * N) / n_threads;
+        unsigned int M_step_end = ((thread_idx + 1) * N) / n_threads;
+        
+        M_step_start = (M_step_start % 8) ? M_step_start + 8 - (M_step_start % 8) : M_step_start;
+        M_step_end = (M_step_end % 8) ? M_step_end + 8 - (M_step_end % 8) : M_step_end;
+        
+        ::ggml_gemv_q4_K_8x8_q8_K(K, (float *)((C) + M_step_start), N,
+                                  (void *)((char *)B + M_step_start * B_step),
+                                  QA.data(), M, M_step_end - M_step_start);
+      }));
+    }
+    
+    for (auto &future : futures) {
+      future.wait();
     }
   } else if (M % 4 != 0) {
-    int n_threads = std::thread::hardware_concurrency();
+    const unsigned int n_threads = std::min(
+      static_cast<unsigned int>(std::thread::hardware_concurrency()), 
+      std::max(1u, N / 32));
     unsigned int blocks_per_4_rows = (K + QK_K - 1) / QK_K;
     unsigned int qa_4_rows_size = sizeof(block_q8_Kx4) * blocks_per_4_rows;
     const size_t qa_row_size = (sizeof(block_q8_K) * K) / QK_K;
@@ -211,41 +233,53 @@ void __ggml_q4_K_8x8_q8_K_GEMM(const unsigned int M, const unsigned int N,
         (QA.data() + (M4 * qa_4_rows_size) + (i - M4 * 4) * qa_row_size), K);
     }
 
-// Compute 4-divisible-M row portion with multithreaded GEMM
-#pragma omp parallel for collapse(1) num_threads(n_threads)
+    // Use BS thread pool for parallel execution
+    std::vector<std::future<void>> futures;
+    futures.reserve(n_threads);
+    
     for (int i = 0; i < n_threads; i++) {
-      unsigned int src0_start = (i * N) / n_threads;
-      unsigned int src0_end = ((i + 1) * N) / n_threads;
+      futures.emplace_back(bspool.submit_task([=, &QA, &B, &C]() {
+        unsigned int src0_start = (i * N) / n_threads;
+        unsigned int src0_end = ((i + 1) * N) / n_threads;
 
-      src0_start =
-        (src0_start % 8) ? src0_start + 8 - (src0_start % 8) : src0_start;
-      src0_end = (src0_end % 8) ? src0_end + 8 - (src0_end % 8) : src0_end;
+        src0_start = (src0_start % 8) ? src0_start + 8 - (src0_start % 8) : src0_start;
+        src0_end = (src0_end % 8) ? src0_end + 8 - (src0_end % 8) : src0_end;
 
-      ::ggml_gemm_q4_K_8x8_q8_K(K, (float *)(C + src0_start), ldc,
-                                (void *)((char *)B + src0_start * B_step),
-                                QA.data(), M4 * 4, src0_end - src0_start);
+        ::ggml_gemm_q4_K_8x8_q8_K(K, (float *)(C + src0_start), ldc,
+                                  (void *)((char *)B + src0_start * B_step),
+                                  QA.data(), M4 * 4, src0_end - src0_start);
+      }));
+    }
+    
+    // Wait for parallel GEMM to complete
+    for (auto &future : futures) {
+      future.wait();
     }
 
     // Compute leftover 1 ~ 3 rows with multithreaded GEMV
-    n_threads = 4;
+    const unsigned int gemv_threads = std::min(4u, std::max(1u, N / 64));
     for (unsigned int pb = M4 * 4; pb < M; pb++) {
-#pragma omp parallel for num_threads(n_threads)
-      for (int thread_idx = 0; thread_idx < n_threads; ++thread_idx) {
-        unsigned int M_step_start = (thread_idx * N) / n_threads; // = 0
-        unsigned int M_step_end =
-          ((thread_idx + 1) * N) / n_threads; // ne01 = N
+      std::vector<std::future<void>> gemv_futures;
+      gemv_futures.reserve(gemv_threads);
+      
+      for (int thread_idx = 0; thread_idx < gemv_threads; ++thread_idx) {
+        gemv_futures.emplace_back(bspool.submit_task([=, &QA, &B, &C]() {
+          unsigned int M_step_start = (thread_idx * N) / gemv_threads;
+          unsigned int M_step_end = ((thread_idx + 1) * N) / gemv_threads;
 
-        M_step_start = (M_step_start % 8)
-                         ? M_step_start + 8 - (M_step_start % 8)
-                         : M_step_start;
-        M_step_end =
-          (M_step_end % 8) ? M_step_end + 8 - (M_step_end % 8) : M_step_end;
+          M_step_start = (M_step_start % 8) ? M_step_start + 8 - (M_step_start % 8) : M_step_start;
+          M_step_end = (M_step_end % 8) ? M_step_end + 8 - (M_step_end % 8) : M_step_end;
 
-        ::ggml_gemv_q4_K_8x8_q8_K(
-          K, (float *)((C + ((pb - M4 * 4) * N) + (M4 * 4 * N)) + M_step_start),
-          N, (void *)((char *)B + M_step_start * B_step),
-          QA.data() + (M4 * qa_4_rows_size) + (pb - M4 * 4) * qa_row_size, 1,
-          M_step_end - M_step_start);
+          ::ggml_gemv_q4_K_8x8_q8_K(
+            K, (float *)((C + ((pb - M4 * 4) * N) + (M4 * 4 * N)) + M_step_start),
+            N, (void *)((char *)B + M_step_start * B_step),
+            QA.data() + (M4 * qa_4_rows_size) + (pb - M4 * 4) * qa_row_size, 1,
+            M_step_end - M_step_start);
+        }));
+      }
+      
+      for (auto &future : gemv_futures) {
+        future.wait();
       }
     }
   } else { // GEMM
@@ -253,8 +287,10 @@ void __ggml_q4_K_8x8_q8_K_GEMM(const unsigned int M, const unsigned int N,
     unsigned int qa_4_rows_size = sizeof(block_q8_Kx4) * blocks_per_4_rows;
     unsigned int M4 = ((M + 3) / 4);
     unsigned int B_step = sizeof(block_q4_K) * (K / QK_K);
-    ///@note OpenMP thread number should be a signed integer
-    int thread_num = std::thread::hardware_concurrency();
+    
+    const unsigned int thread_num = std::min(
+      static_cast<unsigned int>(std::thread::hardware_concurrency()),
+      std::max(1u, N / 32));
 
     unsigned int qa_size = qa_4_rows_size * M4;
     std::vector<char> QA = std::vector<char>(qa_size);
@@ -262,24 +298,31 @@ void __ggml_q4_K_8x8_q8_K_GEMM(const unsigned int M, const unsigned int N,
     // Quantization of activations
     /// @note Heuristic inspection conducted that applying multithreading on
     /// run-time quantization hurts model latency
-    // #pragma omp parallel for collapse(1) num_threads(16)
     for (int i = 0; i < static_cast<int>(M4); i++) {
       ::ggml_quantize_mat_q8_K_4x8(A + 4 * i * K,
                                    QA.data() + i * qa_4_rows_size, K);
     }
 
-#pragma omp parallel for collapse(1) num_threads(thread_num)
+    // Use BS thread pool for parallel GEMM
+    std::vector<std::future<void>> gemm_futures;
+    gemm_futures.reserve(thread_num);
+    
     for (int i = 0; i < thread_num; i++) {
-      unsigned int src0_start = (i * N) / thread_num;
-      unsigned int src0_end = ((i + 1) * N) / thread_num;
+      gemm_futures.emplace_back(bspool.submit_task([=, &QA, &B, &C]() {
+        unsigned int src0_start = (i * N) / thread_num;
+        unsigned int src0_end = ((i + 1) * N) / thread_num;
 
-      src0_start =
-        (src0_start % 8) ? src0_start + 8 - (src0_start % 8) : src0_start;
-      src0_end = (src0_end % 8) ? src0_end + 8 - (src0_end % 8) : src0_end;
+        src0_start = (src0_start % 8) ? src0_start + 8 - (src0_start % 8) : src0_start;
+        src0_end = (src0_end % 8) ? src0_end + 8 - (src0_end % 8) : src0_end;
 
-      ::ggml_gemm_q4_K_8x8_q8_K(K, (float *)(C + src0_start), ldc,
-                                (void *)((char *)B + src0_start * B_step),
-                                QA.data(), M, src0_end - src0_start);
+        ::ggml_gemm_q4_K_8x8_q8_K(K, (float *)(C + src0_start), ldc,
+                                  (void *)((char *)B + src0_start * B_step),
+                                  QA.data(), M, src0_end - src0_start);
+      }));
+    }
+    
+    for (auto &future : gemm_futures) {
+      future.wait();
     }
   }
 }
@@ -323,21 +366,23 @@ void __ggml_gemm_q6_K(const unsigned int M, const unsigned int N,
 
   // GEMV
   if (M == 1) {
+    auto &bspool = ThreadPoolManager::getInstance();
     std::vector<char> quantized_A(A_row_size);
     ::quantize_row_q8_K(A, quantized_A.data(), K);
 
     const void *const quantized_A_data = quantized_A.data();
 
-#pragma omp parallel for collapse(1) num_threads(thread_count)
-    for (int32_t thread_job = 0; thread_job < static_cast<int>(N);
-         thread_job++) {
-      const int32_t B_row_data_offset = B_row_size * thread_job;
-
-      const void *const B_data = (void *)((char *)B + B_row_data_offset);
-
-      ::ggml_vec_dot_q6_K_q8_K(K, &C[thread_job], bs, B_data, bx,
-                               quantized_A_data, by, nrc);
-    }
+    // Use BS thread pool for parallel dot product computation
+    const unsigned int optimal_threads = std::min(16u, std::max(1u, static_cast<unsigned int>(N) / 32));
+    BS::multi_future<void> multi_future = bspool.submit_loop(0, static_cast<int>(N), 
+      [&](int32_t thread_job) {
+        const int32_t B_row_data_offset = B_row_size * thread_job;
+        const void *const B_data = (void *)((char *)B + B_row_data_offset);
+        
+        ::ggml_vec_dot_q6_K_q8_K(K, &C[thread_job], bs, B_data, bx,
+                                 quantized_A_data, by, nrc);
+      });
+    multi_future.wait();
   } else { // GEMM
     const int32_t A_total_size = A_row_size * M;
     std::vector<char> quantized_A(A_total_size);